TestPackageRefactoring - refactoring test package to match source folder. Fix checkstyle issues.
